refactor util scripts
