--- bentoctl_sagemaker/nginx.conf
@@ -1,44 +1,0 @@
-worker_processes 1;
-daemon off; # Prevent forking
-
-pid /tmp/nginx.pid;
-error_log /var/log/nginx/error.log;
-
-events {
-  # defaults
-}
-
-http {
-  include /etc/nginx/mime.types;
-  default_type application/octet-stream;
-  access_log /var/log/nginx/access.log combined;
-
-  upstream gunicorn {
-    server unix:/tmp/gunicorn.sock;
-  }
-
-  server {
-    listen 8080 deferred;
-    client_max_body_size 500m;
-
-    keepalive_timeout 5;
-    proxy_read_timeout 1200s;
-
-    location ~ ^/(ping|invocations) {
-      proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-      proxy_set_header Host $http_host;
-      proxy_redirect off;
-      proxy_pass http://gunicorn;
-    }
-    location /ping {
-      proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-      proxy_set_header Host $http_host;
-      proxy_redirect off;
-      proxy_pass http://gunicorn/healthz;
-    }
-
-    location / {
-      return 404 "{}";
-    }
-  }
-}
