+++ bentoctl_sagemaker/update.py
@@ -5,4 +5,4 @@ def update(bento_path, deployment_name, deployment_spec):
     the cloudformation deploy command can be used to find the changesets and update
     the stack too.
     """
-    deploy(bento_path, deployment_name, deployment_spec)
+    return deploy(bento_path, deployment_name, deployment_spec)
