+++ bentoctl_sagemaker/utils/generate_deployable.py
@@ -1,7 +1,7 @@
 import os
 import shutil
 
-from .utils import is_present, load_bento_tag
+from . import is_present, load_bento_tag
 
 BENTO_SERVICE_SAGEMAKER_DOCKERFILE = """\
 FROM {docker_base_image}
@@ -38,6 +38,7 @@ def generate_sagemaker_target(bento_metadata, bento_path, sagemaker_project_dir)
     if is_present(sagemaker_project_dir):
         return sagemaker_project_dir
 
+    # TODO: get docker tag from bento
     # docker_base_image = bento_metadata.env.docker_base_image
     docker_base_image = "bentoml/bento-server:1.0.0a1-python3.8-debian-runtime"
     shutil.copytree(bento_path, sagemaker_project_dir)
@@ -50,10 +51,10 @@ def generate_sagemaker_target(bento_metadata, bento_path, sagemaker_project_dir)
         )
 
     dir_name = os.path.join(os.path.dirname(__file__))
-    sagemaker_svc_path = os.path.join(dir_name, "./sagemaker_service.py")
+    sagemaker_svc_path = os.path.join(dir_name, "../sagemaker_service.py")
     shutil.copy(sagemaker_svc_path, os.path.join(sagemaker_project_dir, "sagemaker_service.py"))
 
-    serve_file_path = os.path.join(dir_name, "serve")
+    serve_file_path = os.path.join(dir_name, "../serve")
     shutil.copy(serve_file_path, os.path.join(sagemaker_project_dir, "serve"))
 
     # permission 755 is required for entry script 'serve'
@@ -65,7 +66,7 @@ def generate_sagemaker_target(bento_metadata, bento_path, sagemaker_project_dir)
 def generate_deployable(bento_bundle_path, deployment_name):
     bento_tag = load_bento_tag(bento_bundle_path)
 
-    dir_name = f"{bento_tag.name}_{bento_tag.version}_sagemaker_deployable"
+    dir_name = f"{deployment_name}-{bento_tag.name}:{bento_tag.version:4}"
     sagemaker_project_dir = generate_sagemaker_target(
         bento_tag, bento_bundle_path, os.path.abspath(dir_name)
     )
