+++ bentoctl_sagemaker/utils/generate_docker_image_tag.py
